refactor: isolate optional PEF dump diagnostics
